deleted mappers module
